feat: molecule ColorPalette
- 테마컬러 색상표 컴포넌트화
